Ensure fish is sourced once linked
This commit introducts a way to run actions after linking a config.
Mainly as fish config needs to be sourced to work immediately without
manually sourcing or restarting the terminal.

This is now a mix of auto detection when looping over the .config dirs
and explicit scrips as part of the function. This is fine for now but
might consider making all explicit (so one model that defines the config
directory _and_ the post actions) in the future.