Change design to have package modules
This commit changes the approach to a more consistent design. There will
be a conecpt of packages, which essentially implement an interface to
install, link and setup tools. This is all controlled by the package.sh
file and the packages themselves are in the packages dir.

The idea is that each package can add multiple tools it needs, will add
that in future.

The install macos file check executables then install if missing. Note:
could have used the brew env var `HOMEBREW_NO_INSTALL_UPGRADE=1` but
that is much slowe as still needs to update brew and possibly check brew
db etc.